09.09.22

+++ calculations/mathFunctions.h
@@ -1,48 +1,48 @@
 #pragma once
 #include <vector>
 
-int gcd(int a, int b) {
+long long gcd(long long a, long long b) {
     while (b) {
-        int c = a % b;
+        long long c = a % b;
         a = b;
         b = c;
     }
     return a;
 }
 
-int gcd(std::vector<int> a) {
-    int res = 0;
-    for (int i : a) {
+long long gcd(std::vector<long long> a) {
+    long long res = 0;
+    for (long long i : a) {
         res = gcd(i, res);
     }
     return res;
 }
 
-int gcd(const int* a, int n) {
-    std::vector<int> v(n);
-    for (int i = 0; i < n; i++) {
+long long gcd(const long long* a, long long n) {
+    std::vector<long long> v(n);
+    for (long long i = 0; i < n; i++) {
         v[i] = a[i];
     }
     return gcd(v);
 }
 
-int wielandtNumber(int n) {
+long long wielandtNumber(long long n) {
     return n * n - 2 * n + 2;
 }
 
-int slowP(const std::vector<int>& a) {
-    int n = a.size();
-    int N = 0;
-    for (int i = 0; i < n; i++) {
+long long slowP(const std::vector<long long>& a) {
+    long long n = a.size();
+    long long N = 0;
+    for (long long i = 0; i < n; i++) {
         N += a[i];
     }
     N += -n + 1;
-    int ANS = wielandtNumber(N);
+    long long ANS = wielandtNumber(N);
     std::vector<bool> dp(ANS + a[0]);
     dp[0] = true;
-    int cnt = 0;
-    for (int p = 0; p < ANS + a[0]; p++) {
-        for (int i = 0; i < n; i++) {
+    long long cnt = 0;
+    for (long long p = 0; p < ANS + a[0]; p++) {
+        for (long long i = 0; i < n; i++) {
             if (p - a[i] >= 0 && dp[p - a[i]]) {
                 dp[p] = true;
                 break;
@@ -60,35 +60,35 @@ int slowP(const std::vector<int>& a) {
     return -1;
 }
 
-int slowP(const int a[], int n) {
-    std::vector<int> v(n);
-    for (int i = 0; i < n; i++) {
+long long slowP(const long long a[], long long n) {
+    std::vector<long long> v(n);
+    for (long long i = 0; i < n; i++) {
         v[i] = a[i];
     }
     return slowP(v);
 }
 
-int P(std::vector<int> a) {
-    int n = a.size();
-    int N = 0;
-    for (int i = 0; i < n; i++) {
+long long P(std::vector<long long> a) {
+    long long n = a.size();
+    long long N = 0;
+    for (long long i = 0; i < n; i++) {
         N += a[i];
     }
     N -= n + 1;
-    int INF = (a[0] - 1) * (a[n - 1] - 1) + a[0] * (2 * a[n - 1] - 2) + a[0];
-    std::vector<int> M(a[0], INF);
+    long long INF = (a[0] - 1) * (a[n - 1] - 1) + a[0] * (2 * a[n - 1] - 2) + a[0];
+    std::vector<long long> M(a[0], INF);
     M[0] = 0;
-    for (int i = 1; i < n; i++) {
-        for (int k = 0; k < a[0]; k++) {
-            int now = a[i] * k % a[0];
+    for (long long i = 1; i < n; i++) {
+        for (long long k = 0; k < a[0]; k++) {
+            long long now = a[i] * k % a[0];
             M[now] = std::min(M[now], a[i] * k);
         }
     }
-    int deg = 1;
+    long long deg = 1;
     while (deg <= n) {
-        for (int i = 0; i < a[0]; i++) {
-            for (int j = 0; j < a[0]; j++) {
-                int now = (i + j) % a[0];
+        for (long long i = 0; i < a[0]; i++) {
+            for (long long j = 0; j < a[0]; j++) {
+                long long now = (i + j) % a[0];
                 M[now] = std::min(M[now], M[i] + M[j]);
             }
         } 
@@ -97,14 +97,14 @@ int P(std::vector<int> a) {
     return *std::max_element(M.begin(), M.end()) - a[0] + 1;
 }
 
-int P(const int a[], int n) {
-    std::vector<int> v(n);
-    for (int i = 0; i < n; i++) {
+long long P(const long long a[], long long n) {
+    std::vector<long long> v(n);
+    for (long long i = 0; i < n; i++) {
         v[i] = a[i];
     }
     return P(v);
 }
 
-int TforChamomile(const int a[], int n) {
+long long TforChamomile(const long long a[], long long n) {
     return P(a, n) + 2 * a[n - 1] - 2;
 }
